unit_3/project/requirements.txt
# Versioni compatibili e più recenti
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.64
langchain-huggingface==0.2.0
langgraph==0.2.70

# Dataset e ML
datasets==2.21.0
sentence-transformers==3.0.1
numpy==1.26.4
scikit-learn==1.5.0

# Vector store (FAISS - stabile e veloce)
faiss-cpu==1.8.0
langchain-text-splitters==0.3.8

# Tools e utilities
python-dotenv==1.0.1
huggingface_hub>=0.30.2  # Richiesto da langchain-huggingface
duckduckgo-search==6.3.0

# Interfaccia
gradio==5.3.0

# Dipendenze aggiuntive
rank_bm25==0.2.2